+++ packages/fxa-auth-server/test/local/payments/paypal-client.js
@@ -26,6 +26,7 @@ const successfulSetExpressCheckoutResponse = require('./fixtures/paypal/set_expr
 const unSuccessfulSetExpressCheckoutResponse = require('./fixtures/paypal/set_express_checkout_failure.json');
 const successfulDoReferenceTransactionResponse = require('./fixtures/paypal/do_reference_transaction_success.json');
 const unSuccessfulDoReferenceTransactionResponse = require('./fixtures/paypal/do_reference_transaction_failure.json');
+const searchTransactionResponse = require('./fixtures/paypal/transaction_search_success.json');
 const sampleIpnMessage = require('./fixtures/paypal/sample_ipn_message.json')
   .message;
 
@@ -198,11 +199,15 @@ describe('PayPalClient', () => {
       RETURNURL: PLACEHOLDER_URL,
     };
 
+    const defaultOptions = {
+      currencyCode: 'USD',
+    };
+
     it('calls api with correct method and data', () => {
       client.doRequest = sandbox.fake.resolves(
         successfulSetExpressCheckoutResponse
       );
-      client.setExpressCheckout();
+      client.setExpressCheckout(defaultOptions);
       sinon.assert.calledOnceWithExactly(
         client.doRequest,
         'SetExpressCheckout',
@@ -223,7 +228,7 @@ describe('PayPalClient', () => {
         .post(PAYPAL_NVP_ROUTE, client.objectToNVP(expectedPayload))
         .reply(200, client.objectToNVP(unSuccessfulSetExpressCheckoutResponse));
       try {
-        await client.setExpressCheckout();
+        await client.setExpressCheckout(defaultOptions);
         assert.fail('Request should have thrown an error.');
       } catch (err) {
         assert.instanceOf(err, PayPalClientError);
@@ -231,6 +236,24 @@ describe('PayPalClient', () => {
         assert.include(err.raw, 'ACK=Failure');
         assert.equal(err.data.ACK, 'Failure');
       }
+    });
+
+    it('calls api with requested currency', async () => {
+      client.doRequest = sandbox.fake.resolves(
+        successfulSetExpressCheckoutResponse
+      );
+      const currency = 'EUR';
+      await client.setExpressCheckout({
+        currencyCode: currency,
+      });
+      sinon.assert.calledOnceWithExactly(
+        client.doRequest,
+        'SetExpressCheckout',
+        {
+          ...defaultData,
+          PAYMENTREQUEST_0_CURRENCYCODE: currency,
+        }
+      );
     });
   });
 
@@ -364,6 +387,17 @@ describe('PayPalClient', () => {
         .reply(200, 'VERIFIED');
       const result = await client.ipnVerify(sampleIpnMessage);
       assert.equal(result, 'VERIFIED');
+    });
+  });
+
+  describe('transactionSearch', () => {
+    it('calls API with valid message', async () => {
+      client.doRequest = sandbox.fake.resolves(searchTransactionResponse);
+      const response = await client.transactionSearch({
+        startDate: new Date('2010-09-02'),
+        invoice: 'inv-000',
+      });
+      assert.equal(response, searchTransactionResponse);
     });
   });
 });