+++ packages/fxa-auth-server/test/local/payments/paypal.js
@@ -18,6 +18,8 @@ const { mockLog } = require('../../mocks');
 const error = require('../../../lib/error');
 const successfulSetExpressCheckoutResponse = require('./fixtures/paypal/set_express_checkout_success.json');
 const successfulDoReferenceTransactionResponse = require('./fixtures/paypal/do_reference_transaction_success.json');
+const successfulBAUpdateResponse = require('./fixtures/paypal/ba_update_success.json');
+const searchTransactionResponse = require('./fixtures/paypal/transaction_search_success.json');
 const eventCustomerSourceExpiring = require('./fixtures/stripe/event_customer_source_expiring.json');
 const sampleIpnMessage = require('./fixtures/paypal/sample_ipn_message.json');
 const { StripeHelper } = require('../../../lib/payments/stripe');
@@ -91,11 +93,13 @@ describe('PayPalHelper', () => {
   });
 
   describe('getCheckoutToken', () => {
+    const validOptions = { currencyCode: 'USD' };
+
     it('it returns the token from doRequest', async () => {
       paypalHelper.client.doRequest = sinon.fake.resolves(
         successfulSetExpressCheckoutResponse
       );
-      const token = await paypalHelper.getCheckoutToken();
+      const token = await paypalHelper.getCheckoutToken(validOptions);
       assert.equal(token, successfulSetExpressCheckoutResponse.TOKEN);
     });
 
@@ -104,12 +108,24 @@ describe('PayPalHelper', () => {
         new PayPalClientError('Fake', {})
       );
       try {
-        await paypalHelper.getCheckoutToken();
+        await paypalHelper.getCheckoutToken(validOptions);
         assert.fail('Request should have thrown an error.');
       } catch (err) {
         assert.instanceOf(err, PayPalClientError);
         assert.equal(err.name, 'PayPalClientError');
       }
+    });
+
+    it('calls setExpressCheckout with passed options', async () => {
+      paypalHelper.client.setExpressCheckout = sinon.fake.resolves(
+        successfulSetExpressCheckoutResponse
+      );
+      const currencyCode = 'EUR';
+      await paypalHelper.getCheckoutToken({ currencyCode });
+      sinon.assert.calledOnceWithExactly(
+        paypalHelper.client.setExpressCheckout,
+        { currencyCode }
+      );
     });
   });
 
@@ -186,6 +202,75 @@ describe('PayPalHelper', () => {
         assert.instanceOf(err, PayPalClientError);
         assert.equal(err.name, 'PayPalClientError');
       }
+    });
+  });
+
+  describe('cancelBillingAgreement', () => {
+    it('cancels an agreement', async () => {
+      paypalHelper.client.doRequest = sinon.fake.resolves(
+        successfulBAUpdateResponse
+      );
+      const response = await paypalHelper.cancelBillingAgreement('test');
+      assert.isNull(response);
+    });
+
+    it('ignores paypal client errors', async () => {
+      paypalHelper.client.doRequest = sinon.fake.throws(
+        new PayPalClientError('Fake', {})
+      );
+      const response = await paypalHelper.cancelBillingAgreement('test');
+      assert.isNull(response);
+    });
+  });
+
+  describe('searchTransactions', () => {
+    it('returns the data from doRequest', async () => {
+      paypalHelper.client.doRequest = sinon.fake.resolves(
+        searchTransactionResponse
+      );
+      const expectedResponse = [
+        {
+          amount: '5.99',
+          currencyCode: 'USD',
+          email: 'sb-ufoot5037790@personal.example.com',
+          feeAmount: '-0.47',
+          name: 'John Doe',
+          netAmount: '5.52',
+          status: 'Under Review',
+          timestamp: '2021-02-11T17:38:28Z',
+          transactionId: '2TA09271XC591854A',
+          type: 'Payment',
+        },
+        {
+          amount: '5.99',
+          currencyCode: 'USD',
+          email: 'sb-ufoot5037790@personal.example.com',
+          feeAmount: '-0.47',
+          name: 'John Doe',
+          netAmount: '5.52',
+          status: 'Under Review',
+          timestamp: '2021-02-11T17:38:23Z',
+          transactionId: '7WW53923D67853628',
+          type: 'Payment',
+        },
+        {
+          amount: '5.99',
+          currencyCode: 'USD',
+          email: 'sb-ufoot5037790@personal.example.com',
+          feeAmount: '-0.47',
+          name: 'John Doe',
+          netAmount: '5.52',
+          status: 'Under Review',
+          timestamp: '2021-02-11T17:31:05Z',
+          transactionId: '22N88933SF2815829',
+          type: 'Payment',
+        },
+      ];
+      const response = await paypalHelper.searchTransactions({
+        startDate: new Date(),
+        invoice: 'inv-001',
+      });
+      assert.deepEqual(response, expectedResponse);
     });
   });