+++ packages/fxa-payments-server/src/components/PaymentConfirmation/index.stories.tsx
@@ -27,6 +27,7 @@ const selectedPlan: Plan = {
 
 const customer: Customer = {
   billing_name: 'Jane Doe',
+  payment_provider: 'stripe',
   payment_type: 'credit',
   last4: '5309',
   exp_month: '02',