+++ packages/fxa-payments-server/src/store/types.tsx
@@ -73,6 +73,7 @@ export type Customer = {
   exp_month?: string;
   exp_year?: string;
   last4?: string;
+  payment_provider?: string;
   payment_type?: string;
   subscriptions: Array<CustomerSubscription>;
 };