+++ packages/fxa-settings/src/components/DataBlock/index.tsx
@@ -2,52 +2,77 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import React from 'react';
-import { copy } from '../../lib/clipboard';
+import { Localized } from '@fluent/react';
+import React, { useState } from 'react';
+import GetDataTrio from '../GetDataTrio';
+import { Tooltip } from '../Tooltip';
+
+const actionTypeToNotification = {
+  download: 'Downloaded',
+  copy: 'Copied',
+  print: 'Printed',
+} as const;
+
+type actions = keyof typeof actionTypeToNotification;
+type actionFn = (action: actions) => void;
 
 export type DataBlockProps = {
   value: string | string[];
   prefixDataTestId?: string;
-  onCopied?: (copiedValue: string) => void;
+  onAction?: actionFn;
 };
 
 export const DataBlock = ({
   value,
   prefixDataTestId,
-  onCopied,
+  onAction = () => {},
 }: DataBlockProps) => {
   const valueIsArray = Array.isArray(value);
+  const [performedAction, setPerformedAction] = useState<actions>();
+  const dataTestId = prefixDataTestId
+    ? `${prefixDataTestId}-datablock`
+    : 'datablock';
+  const actionCb: actionFn = (action) => {
+    onAction(action);
 
-  async function copyToClipboard() {
-    const copyValue = valueIsArray
-      ? (value as string[]).join(', ')
-      : (value as string);
-
-    await copy(copyValue);
-    onCopied && onCopied(copyValue);
-  }
-  function formatDataTestId(id: string) {
-    return prefixDataTestId ? `${prefixDataTestId}-${id}` : id;
-  }
+    if (actionTypeToNotification[action]) {
+      setPerformedAction(action);
+    }
+  };
 
   return (
-    <button
-      className={`flex rounded-xl px-7 font-mono text-sm text-green-900 bg-green-800 bg-opacity-10 flex-wrap hover:bg-opacity-20 focus:bg-opacity-30 active:bg-opacity-30 ${
-        valueIsArray ? 'py-4' : 'py-5'
-      }`}
-      data-testid={formatDataTestId('datablock-button')}
-      onClick={copyToClipboard}
-    >
-      {valueIsArray ? (
-        (value as string[]).map((item) => (
-          <span key={item} className="flex-50% py-1">
-            {item}
-          </span>
-        ))
-      ) : (
-        <span>{value}</span>
-      )}
-    </button>
+    <>
+      <div
+        className={`flex rounded-xl px-7 font-mono text-center text-sm text-green-900 bg-green-800 bg-opacity-10 flex-wrap relative mb-8 ${
+          valueIsArray ? 'py-4' : 'py-5'
+        }`}
+        data-testid={dataTestId}
+      >
+        {valueIsArray ? (
+          (value as string[]).map((item) => (
+            <span key={item} className="flex-50% py-1">
+              {item}
+            </span>
+          ))
+        ) : (
+          <span>{value}</span>
+        )}
+        {performedAction && (
+          <Localized
+            id={`datablock-${performedAction}`}
+            attrs={{ message: true }}
+          >
+            <Tooltip
+              prefixDataTestId={`datablock-${performedAction}`}
+              message={actionTypeToNotification[performedAction]}
+              position="bottom"
+              className="mt-1"
+            ></Tooltip>
+          </Localized>
+        )}
+      </div>
+      <GetDataTrio {...{ value, onAction: actionCb }} />
+    </>
   );
 };