+++ packages/fxa-settings/src/components/Page2faReplaceRecoveryCodes/index.tsx
@@ -8,7 +8,6 @@ import React, { useEffect, useState } from 'react';
 import FlowContainer from '../FlowContainer';
 import VerifiedSessionGuard from '../VerifiedSessionGuard';
 import DataBlock from '../DataBlock';
-import GetDataTrio from '../GetDataTrio';
 import { useSession } from '../../models';
 import { useAlertBar, useMutation } from '../../lib/hooks';
 import { AlertBar } from '../AlertBar';
@@ -62,12 +61,11 @@ export const Page2faReplaceRecoveryCodes = (_: RouteComponentProps) => {
       <div className="my-2" data-testid="2fa-recovery-codes">
         <Localized id="tfa-replace-code-success">
           New codes have been created. Save these one-time use codes in a safe
-          place—you’ll need them to access your account if you don’t have your
+          place — you’ll need them to access your account if you don’t have your
           mobile device.
         </Localized>
         <div className="mt-6 flex flex-col items-center h-40 justify-between">
           <DataBlock value={recoveryCodes}></DataBlock>
-          <GetDataTrio value={recoveryCodes}></GetDataTrio>
         </div>
       </div>
       <div className="flex justify-center mt-6 mb-4 mx-auto max-w-64">