+++ packages/fxa-settings/src/components/PageTwoStepAuthentication/index.tsx
@@ -13,7 +13,6 @@ import React, { useEffect, useState } from 'react';
 import VerifiedSessionGuard from '../VerifiedSessionGuard';
 import AlertBar from '../AlertBar';
 import DataBlock from '../DataBlock';
-import GetDataTrio from '../GetDataTrio';
 import { useSession } from '../../models';
 import { checkCode, getCode } from '../../lib/totp';
 import { HomePath } from '../../constants';
@@ -192,7 +191,7 @@ export const PageTwoStepAuthentication = (_: RouteComponentProps) => {
           l10n.getString(
             'tfa-cannot-verify-code',
             null,
-            'There was a problem verifiying your recovery code.'
+            'There was a problem verifying your recovery code.'
           )
         );
       }
@@ -240,7 +239,7 @@ export const PageTwoStepAuthentication = (_: RouteComponentProps) => {
 
   return (
     <FlowContainer
-      title={l10n.getString('tfa-title')}
+      title={l10n.getString('tfa-title', null, 'Two-Step Authentication')}
       {...{ subtitle, onBackButtonClick: moveBack }}
     >
       {alertBar.visible && (
@@ -370,11 +369,10 @@ export const PageTwoStepAuthentication = (_: RouteComponentProps) => {
               have your mobile device.
             </Localized>
             <div className="mt-6 flex flex-col items-center h-40 justify-between">
-              <DataBlock value={recoveryCodes}></DataBlock>
-              <GetDataTrio
+              <DataBlock
                 value={recoveryCodes}
                 onAction={logDataTrioActionEvent}
-              ></GetDataTrio>
+              ></DataBlock>
             </div>
           </div>
           <div className="flex justify-center mt-6 mb-4 mx-auto max-w-64">