+++ packages/fxa-geodb/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history
 
+## 1.200.0
+
+No changes.
+
 ## 1.199.0
 
 No changes.